+++ src/hover.ts
@@ -1,8 +1,9 @@
 import {hoverTooltip} from '@codemirror/view';
 import {loadScript} from '@bhsd/common';
 import type {Tooltip, TooltipView, EditorView} from '@codemirror/view';
+import type {Text} from '@codemirror/state';
 import type {LanguageServiceBase} from 'wikiparser-node/extensions/typings';
-import type {MarkupContent} from 'vscode-languageserver-types';
+import type {MarkupContent, Position} from 'vscode-languageserver-types';
 import type * as MarkdownIt from 'markdown-it';
 
 declare const markdownit: () => MarkdownIt;
@@ -14,7 +15,9 @@ let md: MarkdownIt | undefined;
  * 获取当前编辑器的语言服务
  * @param view EditorView 实例
  */
-const getLSP = (view: EditorView): LanguageServiceBase | undefined => {
+export const getLSP = (view: EditorView): LanguageServiceBase | undefined => {
+	void loadScript('npm/wikiparser-node/extensions/dist/base.min.js', 'wikiparse');
+	void loadScript('npm/wikiparser-node/extensions/dist/lsp.min.js', 'wikiparse.LanguageService');
 	if (!('wikiparse' in globalThis && wikiparse.LanguageService)) {
 		return undefined;
 	} else if (lsps.has(view)) {
@@ -25,11 +28,20 @@ const getLSP = (view: EditorView): LanguageServiceBase | undefined => {
 	return lsp;
 };
 
+/**
+ * 将索引转换为位置
+ * @param doc Text 实例
+ * @param index 索引
+ */
+export const indexToPos = (doc: Text, index: number): Position => {
+	const line = doc.lineAt(index);
+	return {line: line.number - 1, character: index - line.from};
+};
+
 export default hoverTooltip(async (view, pos): Promise<Tooltip | null> => {
 	const {state: {doc}} = view,
-		line = doc.lineAt(pos),
 		hover = await getLSP(view)
-			?.provideHover(view.state.doc.toString(), {line: line.number - 1, character: pos - line.from});
+			?.provideHover(doc.toString(), indexToPos(doc, pos));
 	if (hover) {
 		await loadScript('npm/markdown-it/dist/markdown-it.min.js', 'markdownit', true);
 		md ??= markdownit();